options: require "pprof" build tag to enable profiling
