+++ src/options.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"os"
 	"regexp"
-	"runtime"
-	"runtime/pprof"
 	"strconv"
 	"strings"
 	"unicode"
@@ -2264,66 +2262,6 @@ func postProcessOptions(opts *Options) {
 	if opts.Scheme != "default" {
 		processScheme(opts)
 	}
-}
-
-func (o *Options) initProfiling() error {
-	if o.CPUProfile != "" {
-		f, err := os.Create(o.CPUProfile)
-		if err != nil {
-			return fmt.Errorf("could not create CPU profile: %w", err)
-		}
-
-		if err := pprof.StartCPUProfile(f); err != nil {
-			return fmt.Errorf("could not start CPU profile: %w", err)
-		}
-
-		util.AtExit(func() {
-			pprof.StopCPUProfile()
-			if err := f.Close(); err != nil {
-				fmt.Fprintln(os.Stderr, "Error: closing cpu profile:", err)
-			}
-		})
-	}
-
-	stopProfile := func(name string, f *os.File) {
-		if err := pprof.Lookup(name).WriteTo(f, 0); err != nil {
-			fmt.Fprintf(os.Stderr, "Error: could not write %s profile: %v\n", name, err)
-		}
-		if err := f.Close(); err != nil {
-			fmt.Fprintf(os.Stderr, "Error: closing %s profile: %v\n", name, err)
-		}
-	}
-
-	if o.MEMProfile != "" {
-		f, err := os.Create(o.MEMProfile)
-		if err != nil {
-			return fmt.Errorf("could not create MEM profile: %w", err)
-		}
-		util.AtExit(func() {
-			runtime.GC()
-			util.AtExit(func() { stopProfile("allocs", f) })
-		})
-	}
-
-	if o.BlockProfile != "" {
-		runtime.SetBlockProfileRate(1)
-		f, err := os.Create(o.BlockProfile)
-		if err != nil {
-			return fmt.Errorf("could not create BLOCK profile: %w", err)
-		}
-		util.AtExit(func() { stopProfile("block", f) })
-	}
-
-	if o.MutexProfile != "" {
-		runtime.SetMutexProfileFraction(1)
-		f, err := os.Create(o.MutexProfile)
-		if err != nil {
-			return fmt.Errorf("could not create MUTEX profile: %w", err)
-		}
-		util.AtExit(func() { stopProfile("mutex", f) })
-	}
-
-	return nil
 }
 
 func expectsArbitraryString(opt string) bool {
